implementing spline graph in the altitude adjust dialog
git-svn-id: https://mytourbook.svn.sourceforge.net/svnroot/mytourbook/trunk@979 5a91af82-4d22-0410-8bcb-a41d7c89129f

+++ net.tourbook/src/net/tourbook/data/TourData.java
@@ -604,7 +604,7 @@ public class TourData implements Comparable<Object> {
 	 * Contains a spline data serie
 	 */
 	@Transient
-	public int[]						dataSerieSpline;
+	public float[]						dataSerieSpline;
 
 	/**
 	 * when a tour is not saved, the tour id is not defined, therefore the tour data are provided
